autogen/pyproject.toml
[project]
name = "spawnhub-example-autogen"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "pyautogen>=0.2.36",
    "opentelemetry-sdk>=1.25.0",
    "opentelemetry-exporter-otlp-proto-http>=1.25.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = []
